biên bản hao dôi bdg

+++ src/main/java/com/tcdt/qlnvhang/request/xuathang/daugia/xuatkho/SearchXhDgBbTinhKho.java
@@ -9,7 +9,7 @@ import java.time.LocalDate;
 public class SearchXhDgBbTinhKho extends BaseRequest {
   private String maDvi;
   private String dvql;
-  private Long nam;
+  private Integer nam;
   private String soQdGiaoNvXh;
   private String soBbTinhKho;
   private LocalDate ngayBatDauXuatTu;
